fix: learning module convention audit - split oversized components, MediatR controllers, AsNoTracking queries, admin-service domain split, useApiMutation
