+++ Backend/EY.HRPlatform.Training/Features/Admin/Queries/GetCompletionByGradeQueryHandler.cs
@@ -18,11 +18,13 @@ public class GetCompletionByGradeQueryHandler
         GetCompletionByGradeQuery request, CancellationToken cancellationToken)
     {
         var grades = await _db.Grades
+            .AsNoTracking()
             .OrderBy(g => g.Level)
             .ToListAsync(cancellationToken);
 
         // Batch 1: all profiles with grade + service line
         var profiles = await _db.EmployeeProfiles
+            .AsNoTracking()
             .Where(ep => ep.GradeId != null && ep.ServiceLineId != null)
             .Select(ep => new
             {
@@ -34,6 +36,7 @@ public class GetCompletionByGradeQueryHandler
 
         // Batch 2: all curriculum mappings
         var mappings = await _db.CurriculumMappings
+            .AsNoTracking()
             .Select(m => new { m.GradeId, m.ServiceLineId, m.TrainingId })
             .ToListAsync(cancellationToken);
 
@@ -48,6 +51,7 @@ public class GetCompletionByGradeQueryHandler
 
         var completedProgressRaw = profiledEmployeeIds.Count > 0 && allCurriculumTrainingIds.Count > 0
             ? await _db.TrainingProgress
+                .AsNoTracking()
                 .Where(tp =>
                     profiledEmployeeIds.Contains(tp.EmployeeId)
                     && allCurriculumTrainingIds.Contains(tp.TrainingId)
