+++ Backend/EY.HRPlatform.Training/Models/Requests/ProvisionEmployeeRequest.cs
@@ -1,0 +1,4 @@
+namespace EY.HRPlatform.Training.Models.Requests;
+
+/// <summary>Request body for the employee provision endpoint.</summary>
+public sealed record ProvisionEmployeeRequest(Guid EmployeeId);
