+++ Frontend/apps/learning/src/components/training-detail-page.tsx
@@ -1,196 +1,61 @@
 "use client";
 
-import { useCallback } from "react";
-import { useRouter } from "next/navigation";
-import { Button } from "@repo/ui";
-import { useApiQuery } from "@repo/api/react";
-import { CalendarDays, ChevronRight, AlertTriangle, Award, Loader2, Play } from "lucide-react";
+import { CalendarDays, AlertTriangle, Award } from "lucide-react";
 import type { TrainingDetailPageProps } from "@/types/component-props";
 import { CATEGORY_CONFIG, LEVEL_CONFIG } from "@/data/categories";
 import { BADGE_LEVEL_CONFIG } from "@/data/badge-config";
-import { useEnroll } from "@/hooks/use-enroll";
-import { getEnrollmentStatus } from "@/services/learning-service";
 import { PageHeader } from "./page-header";
 import { PageBreadcrumb } from "./page-breadcrumb";
-import {
-  TrainingStatsGrid,
-  ChapterList,
-  ExamSection,
-  InstructorCard,
-  TrainingTagsCard,
-  OnSiteCoursesList,
-} from "./training-detail";
+import { TrainingStatsGrid, ChapterList, ExamSection, InstructorCard, TrainingTagsCard, OnSiteCoursesList } from "./training-detail";
+import { TrainingEnrollCta } from "./training-enroll-cta";
 
 export function TrainingDetailPage({ training }: TrainingDetailPageProps) {
-  const router = useRouter();
   const category = CATEGORY_CONFIG[training.category];
   const level = LEVEL_CONFIG[training.level];
   const badge = BADGE_LEVEL_CONFIG[training.badgeLevel];
-  const { handleEnroll, isLoading, enrolled } = useEnroll(training.id);
-
-  const fetchEnrollment = useCallback(
-    () => getEnrollmentStatus(training.id),
-    [training.id],
-  );
-
-  const { data: enrollment, isLoading: checkingEnrollment } = useApiQuery(
-    fetchEnrollment,
-  );
-
-  const isEnrolled = enrolled || !!enrollment;
-
-  const handleContinueLearning = () => {
-    router.push(`/training/${encodeURIComponent(training.id)}/learn`);
-  };
 
   return (
     <div className="min-h-full">
-      <PageBreadcrumb
-        backHref="/"
-        backLabel="Back"
-        items={[
-          { label: "Catalog", href: "/" },
-          { label: training.title },
-        ]}
-      />
-      <PageHeader
-        moduleTitle="Training Details"
-        title={training.title}
-        description={training.description}
-      >
-        {/* Badges row */}
-        <div
-          className="ey-animate-fade-up mt-4 flex items-center gap-3 flex-wrap"
-          style={{ animationDelay: "200ms" }}
-        >
-          <span
-            className={`inline-flex items-center rounded-full border px-3 py-1 text-xs font-semibold tracking-wide uppercase ${category.badgeClass}`}
-          >
-            {category.label}
-          </span>
-          <span className="flex items-center gap-1.5 text-sm text-muted-foreground">
-            <span className={`h-2 w-2 rounded-full ${level.dotClass}`} />
-            {level.label}
-          </span>
-          <span
-            className={`inline-flex items-center gap-1.5 rounded-full border px-2.5 py-0.5 text-xs font-semibold ${badge.className}`}
-          >
-            <Award className="h-3 w-3" aria-hidden="true" />
-            {badge.label}
-          </span>
+      <PageBreadcrumb backHref="/" backLabel="Back" items={[{ label: "Catalog", href: "/" }, { label: training.title }]} />
+      <PageHeader moduleTitle="Training Details" title={training.title} description={training.description}>
+        <div className="ey-animate-fade-up mt-4 flex items-center gap-3 flex-wrap" style={{ animationDelay: "200ms" }}>
+          <span className={`inline-flex items-center rounded-full border px-3 py-1 text-xs font-semibold tracking-wide uppercase ${category.badgeClass}`}>{category.label}</span>
+          <span className="flex items-center gap-1.5 text-sm text-muted-foreground"><span className={`h-2 w-2 rounded-full ${level.dotClass}`} />{level.label}</span>
+          <span className={`inline-flex items-center gap-1.5 rounded-full border px-2.5 py-0.5 text-xs font-semibold ${badge.className}`}><Award className="h-3 w-3" aria-hidden="true" />{badge.label}</span>
           {training.isMandatory && (
-            <span className="inline-flex items-center gap-1 rounded-full bg-amber-50 border border-amber-200 px-2.5 py-0.5 text-xs font-semibold text-amber-700">
-              <AlertTriangle className="h-3 w-3" aria-hidden="true" />
-              Mandatory
-            </span>
+            <span className="inline-flex items-center gap-1 rounded-full bg-amber-50 border border-amber-200 px-2.5 py-0.5 text-xs font-semibold text-amber-700"><AlertTriangle className="h-3 w-3" aria-hidden="true" />Mandatory</span>
           )}
           {training.trainingType === "OnSite" && (
-            <span className="inline-flex items-center rounded-full bg-blue-50 border border-blue-200 px-2.5 py-0.5 text-xs font-semibold text-blue-700">
-              On-Site Training
-            </span>
+            <span className="inline-flex items-center rounded-full bg-blue-50 border border-blue-200 px-2.5 py-0.5 text-xs font-semibold text-blue-700">On-Site Training</span>
           )}
         </div>
       </PageHeader>
 
-      {/* ── Body ── */}
       <div className="px-8 py-8">
         <div className="grid gap-8 lg:grid-cols-[1fr_340px]">
-          {/* ── Left column ── */}
           <div className="space-y-8">
             <TrainingStatsGrid training={training} />
             {training.trainingType === "OnSite" ? (
-              <OnSiteCoursesList
-                courses={training.onSiteCourses ?? []}
-                scheduledDate={training.scheduledDate}
-              />
+              <OnSiteCoursesList courses={training.onSiteCourses ?? []} scheduledDate={training.scheduledDate} />
             ) : (
               <>
-                <ChapterList
-                  chapters={training.chapters}
-                  chaptersCount={training.chaptersCount}
-                />
-                <ExamSection
-                  exam={training.exam}
-                  chaptersCount={training.chaptersCount}
-                />
+                <ChapterList chapters={training.chapters} chaptersCount={training.chaptersCount} />
+                <ExamSection exam={training.exam} chaptersCount={training.chaptersCount} />
               </>
             )}
           </div>
 
-          {/* ── Right column / Sidebar ── */}
           <aside className="space-y-6">
-            <InstructorCard
-              name={training.instructor}
-              role={training.instructorRole}
-            />
+            <InstructorCard name={training.instructor} role={training.instructorRole} />
             <TrainingTagsCard tags={training.tags} />
-
-            {/* Updated date */}
-            <div
-              className="ey-animate-fade-up rounded-2xl border border-border/50 bg-white p-6"
-              style={{ animationDelay: "300ms" }}
-            >
+            <div className="ey-animate-fade-up rounded-2xl border border-border/50 bg-white p-6" style={{ animationDelay: "300ms" }}>
               <div className="flex items-center gap-2 text-sm text-muted-foreground">
                 <CalendarDays className="h-4 w-4" aria-hidden="true" />
-                <span>
-                  Last updated{" "}
-                  {new Date(
-                    training.updatedAt + "T00:00:00"
-                  ).toLocaleDateString("en-US", {
-                    month: "long",
-                    day: "numeric",
-                    year: "numeric",
-                  })}
-                </span>
+                <span>Last updated {new Date(training.updatedAt + "T00:00:00").toLocaleDateString("en-US", { month: "long", day: "numeric", year: "numeric" })}</span>
               </div>
             </div>
-
-            {/* CTA */}
-            <div
-              className="ey-animate-fade-up sticky top-6"
-              style={{ animationDelay: "350ms" }}
-            >
-              {checkingEnrollment ? (
-                <Button
-                  disabled
-                  className="w-full h-12 text-sm font-semibold"
-                >
-                  <Loader2 className="h-4 w-4 animate-spin" aria-hidden="true" />
-                </Button>
-              ) : isEnrolled ? (
-                <Button
-                  onClick={handleContinueLearning}
-                  className="w-full bg-emerald-600 hover:bg-emerald-700 text-white gap-2 shadow-lg transition-all hover:shadow-xl hover:gap-3 h-12 text-sm font-semibold"
-                >
-                  <Play className="h-4 w-4" aria-hidden="true" />
-                  {training.trainingType === "OnSite" ? "Access Courses" : "Continue Learning"}
-                  <ChevronRight
-                    className="h-4 w-4 transition-transform"
-                    aria-hidden="true"
-                  />
-                </Button>
-              ) : (
-                <Button
-                  onClick={handleEnroll}
-                  disabled={isLoading}
-                  className="w-full ey-bg-dark hover:ey-bg-dark-deep text-white gap-2 shadow-lg transition-all hover:shadow-xl hover:gap-3 h-12 text-sm font-semibold"
-                >
-                  {isLoading ? (
-                    <>
-                      <Loader2 className="h-4 w-4 animate-spin" aria-hidden="true" />
-                      Enrolling...
-                    </>
-                  ) : (
-                    <>
-                      Enroll Now
-                      <ChevronRight
-                        className="h-4 w-4 transition-transform"
-                        aria-hidden="true"
-                      />
-                    </>
-                  )}
-                </Button>
-              )}
+            <div className="ey-animate-fade-up sticky top-6" style={{ animationDelay: "350ms" }}>
+              <TrainingEnrollCta trainingId={training.id} trainingType={training.trainingType} />
             </div>
           </aside>
         </div>
