+++ Frontend/apps/learning/src/hooks/index.ts
@@ -1,2 +1,19 @@
-// Custom hooks for the Learning module
-// Export hooks from this file as they are created
+export { useAdminDashboardData } from "./use-admin-dashboard-data";
+export { useChapterBuilder } from "./use-chapter-builder";
+export { useChapterForm } from "./use-chapter-form";
+export { useCoursePlayer } from "./use-course-player";
+export { useDashboardData } from "./use-dashboard-data";
+export { useEditTrainingWizard } from "./use-edit-training-wizard";
+export { useEnroll } from "./use-enroll";
+export { useExamBuilder } from "./use-exam-builder";
+export { useExamPlayer } from "./use-exam-player";
+export {
+  useProgrammeMatrix,
+  useCompletionByGrade,
+  useCompletionByServiceLine,
+  useCompletionTrend,
+  useCellEmployees,
+} from "./use-programme-dashboard";
+export { useTrainingForm } from "./use-training-form";
+export { useTrainingWizard } from "./use-training-wizard";
+export { useChapterTitle } from "./use-chapter-title";
