+++ Frontend/apps/learning/src/services/index.ts
@@ -6,3 +6,5 @@ export {
   enrollInTraining,
   updateChapterProgress,
 } from "./learning-service";
+
+export * from "./admin-service";
